Remove unnecessary dexdeps-docs and dexdep-doc-stubs
The only purpose of these seemed to be documenting the API of the
implementation classes of the `dexdeps` tool. That is not the purpose
for which `droidstubs` was intended and it is blocking improvements to
the build's handling of flagged APIs so this change removed them.

Flag: EXEMPT build change to improve flag support
Bug: 425309457
Test: TH
Change-Id: I0a4ca14a243d9d46cce45b3e3340581aa3dc32b0